Move track/3 to Session
